Fix shift into sign

+++ code/client/cl_cin.c
@@ -601,7 +601,7 @@ static unsigned int yuv_to_rgb24( long y, long u, long v )
 	if (r<0) r = 0; if (g<0) g = 0; if (b<0) b = 0;
 	if (r > 255) r = 255; if (g > 255) g = 255; if (b > 255) b = 255;
 	
-	return LittleLong ((r)|(g<<8)|(b<<16)|(255<<24));
+	return LittleLong ((unsigned long)((r)|(g<<8)|(b<<16))|(255UL<<24));
 }
 
 /******************************************************************************
